requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
langchain-community
sentence-transformers
faiss-cpu
pypdf
tavily-python
python-dotenv
groq